change indexer to prepare webpages in batches
