+++ crates/core/src/webpage/html/mod.rs
@@ -493,7 +493,6 @@ mod tests {
     use std::collections::HashMap;
 
     use crate::{
-        rake::RakeModel,
         schema::create_schema,
         webpage::{url_ext::UrlExt, Link},
     };
@@ -741,9 +740,7 @@ mod tests {
         assert!(!webpage.all_text().unwrap().is_empty());
 
         let schema = create_schema();
-        webpage
-            .into_tantivy(&schema, &RakeModel::default())
-            .unwrap();
+        webpage.as_tantivy(&schema).unwrap();
     }
 
     #[test]
